# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_checkpoint_server_internal_models.Game:
    properties:
      cover_url:
        type: string
      created_at:
        type: string
      id:
        type: string
      igdb_id:
        type: integer
      release_date:
        type: string
      steam_app_id:
        description: Used to map imported games
        type: integer
      summary:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_checkpoint_server_internal_models.Review:
    properties:
      contains_spoiler:
        type: boolean
      content:
        type: string
      created_at:
        type: string
      game:
        $ref: '#/definitions/github_com_checkpoint_server_internal_models.Game'
      game_id:
        type: string
      id:
        type: string
      rating:
        description: 0-5 stars
        type: number
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_checkpoint_server_internal_models.User'
      user_id:
        type: string
    type: object
  github_com_checkpoint_server_internal_models.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      steam_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  internal_handlers.CreateReviewRequest:
    properties:
      contains_spoiler:
        type: boolean
      content:
        type: string
      igdb_id:
        type: integer
      rating:
        description: 0-5
        type: number
      title:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: The Game Journal (Letterboxd for games) API
  title: Checkpoint MVP API
  version: "1.0"
paths:
...
  /games/{id}:
    get:
      consumes:
      - application/json
      description: Fetch detailed information for a specific game by its IGDB ID
      parameters:
      - description: IGDB Game ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Game by ID
      tags:
      - games
  /games/discover:
    get:
      consumes:
      - application/json
      description: Fetch popular mainstream games (All time)
      parameters:
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Discover Games
      tags:
      - games
  /games/new-releases:
    get:
      consumes:
      - application/json
      description: Fetch newly released games
      parameters:
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: New Releases
      tags:
      - games
  /games/search:
    get:
      consumes:
      - application/json
      description: Search for games on IGDB
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search Games
      tags:
      - games
  /games/trending:
    get:
      consumes:
      - application/json
      description: Fetch currently trending games
      parameters:
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Trending Games
      tags:
      - games
...
  /reviews:
    post:
      consumes:
      - application/json
      description: Write a review and rate a game
      parameters:
      - description: Create Review Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_checkpoint_server_internal_models.Review'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a Review
      tags:
      - reviews
  /reviews/game/{igdb_id}:
    get:
      consumes:
      - application/json
      description: Get all reviews for a specific game by IGDB ID
      parameters:
      - description: IGDB ID
        in: path
        name: igdb_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_checkpoint_server_internal_models.Review'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Game Reviews
      tags:
      - reviews
  /reviews/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Get all reviews made by a specific user
      parameters:
      - description: User UUID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_checkpoint_server_internal_models.Review'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get User Reviews
      tags:
      - reviews
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"